refactor: 축제 이름 컴포넌트 분리
