+++ src/components/common/BorderCardComponent.tsx
@@ -1,0 +1,5 @@
+const BorderCardComponent = ({ children }: { children: React.ReactNode }) => {
+  return <div className="p-4 border-white rounded-lg shadow-sm">{children}</div>;
+};
+
+export default BorderCardComponent;
